+++ doc/changelog.md
@@ -9,6 +9,7 @@
 - [S] remove Site Title animation.
   移除 Title 动画。
 - [S] Banner height -> 550px -> 350px (PC)
+- [F] fix topView post order.
 
 ## 2019/11/2